requirements.txt
SpeechRecognition==3.10.0
requests==2.31.0
python-dotenv==1.0.0
selenium==4.15.2
webdriver_manager==4.0.1
PyAudio==0.2.13